% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copula.R
\name{estimate_negative_binomial_copula}
\alias{estimate_negative_binomial_copula}
\alias{estimate_zi_poisson_copula}
\alias{estimate_poisson_copula}
\alias{estimate_gaussian_copula}
\title{todo}
\usage{
estimate_zi_poisson_copula(data_matrix, cores)

estimate_negative_binomial_copula(data_matrix, cores)

estimate_poisson_copula(data_matrix, cores)

estimate_gaussian_copula(data_matrix, cores)
}
\arguments{
\item{data_matrix}{The data to estimate parameters from.}

\item{cores}{The number of CPU cores to use in estimation by scDesign3.}
}
\value{
todo

todo

todo

todo
}
\description{
Given data, computes todo

Given data, computes todo

Given data, computes todo

Given data, computes todo
}
