% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recall.R
\name{get_seurat_obj_with_artificial_variables}
\alias{get_seurat_obj_with_artificial_variables}
\title{Returns a Seurat object that contains additional (fake) RNA
expression counts.}
\usage{
get_seurat_obj_with_artificial_variables(
  seurat_obj,
  assay = "RNA",
  null_method = "ZIP",
  verbose = TRUE,
  cores
)
}
\arguments{
\item{seurat_obj}{A Seurat object containing RNA expression counts.}

\item{assay}{The assay to generate artificial variables from.}

\item{null_method}{The generating distribution for the synthetic null variables (ZIP, NB, ZIP-copula, NB-copula)}

\item{verbose}{Whether or not to show logging.}

\item{cores}{The number of cores to use in generating synthetic null variables.}
}
\value{
A Seurat object that contains the original variable features and an
equal number of artificial features.
}
\description{
Given a Seurat object, returns a new Seurat object whose RNA
expression counts includes the
variable features from the original object and an equal number of artificial
features.
}
